requirements.txt
anthropic>=0.40.0
PyMuPDF>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
click>=8.1.0
tqdm>=4.66.0
streamlit>=1.30.0
python-docx>=1.1.0
boto3>=1.34.0
